将写入速率限制的类型从 float 更改为 double，以提高精度

+++ cpp/lsm/core/db.h
@@ -15,7 +15,7 @@ namespace lsm {
 
 struct Options {
     bool sync = false; // true: fsync on every write, false: rely on background sync
-    float write_rate_limit = 0.0; // bytes/sec, 0.0 means no limit
+    double write_rate_limit = 0.0; // bytes/sec, 0.0 means no limit
 };
 
 class DB {
